Removed "fail safe" timer from audio output device as it basically doesn't help, and actually makes it slightly worse.
This means that old style "multimedia" interface for sound playback is useless on Vista and later due to it's emulation over ASIO causing latency and "jitter" in the ply back, jitter occurring on both side s of the jitter buffer does not work and we get bad audio playback as a result. Essentially, for Vista and later, use DirectSound.